Sacado + Kokkos Issue #566
Synchronous changes in Sacado to mirror those in Kokkos for the
CommonViewAllocProp ViewCtorProp functionality newly added to help find
common value_type among views and store the fad_dim if a Fad type is
present.
Changes to the internal implementation of the DynRankView were made
(createLayout, computeRank) and corresponding changes to specializations
of those functions in Sacado are necessary for next Kokkos snapshot
(2.03.06) into Trilinos.